Rename Scheduler to Queue
